Classmerger: change default `strict_merged_args`
Default `strict_merged_args` value is set to True, so that a `TypeError`
exception is raised by default if positional arguments are missing when
invoking constructor `__init__` or other methods listed in `invoke_all`.
If `strict_merged_args` is set to False, then the methods are silently
skipped in case of missing positional arguments.

Signed-off-by: Patrizio Gelosi <patrizio.gelosi@amarulasolutions.com>